Unificação do cadastro em uma unica pagina e controller
