+++ laravel/routes/web.php
@@ -15,9 +15,6 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/teste', function(){ return view('teste.index');});
 
-Route::get('/cadastro', function(){ return view('cadastro.index');});
-Route::post('/cadastro/cliente', 'App\Http\Controllers\CadastroClienteController@index');
-Route::post('/cadastro/cliente/criar', 'App\Http\Controllers\CadastroClienteController@store');
-Route::post('/cadastro/profissional', 'App\Http\Controllers\CadastroProfissionalController@index');
-Route::post('/cadastro/profissional/criar', 'App\Http\Controllers\CadastroProfissionalController@store');
+Route::get('/cadastro', 'App\Http\Controllers\CadastroController@index');
+Route::post('/cadastro', 'App\Http\Controllers\CadastroController@store');
 Route::get('/', function(){ return view('home.index');});
